refactor: restructure entry point for web app
