+++ README.md
@@ -12,7 +12,7 @@ In order to run the application locally, you need:
 - [Docker Compose](https://docs.docker.com/compose/)
 - Make sure you have Python installed. version 3.10 or higher is required.
 
-Building the Dockerfile requires access to the private Github repository (TODO: Link to fortran repo here).  
+Building the Dockerfile requires access to the private Github repository with the fortan library source (https://github.com/equinor/gpa-libhg).  
 The simplest way to manage this is by adding a __deploy key__ to the repository.  
 Generate the keypair like so;
