--- web/src/authConfig.ts
@@ -1,8 +1,0 @@
-export const authConfig = {
-  clientId: process.env.REACT_APP_AUTH_CLIENT_ID || '',
-  authorizationEndpoint: process.env.REACT_APP_AUTH_ENDPOINT || '',
-  tokenEndpoint: process.env.REACT_APP_TOKEN_ENDPOINT || '',
-  scope: process.env.REACT_APP_AUTH_SCOPE || '',
-  redirectUri: process.env.REACT_APP_AUTH_REDIRECT_URI || '',
-  logoutEndpoint: process.env.REACT_APP_LOGOUT_ENDPOINT || '',
-}
